requirements.txt
astroid==2.3.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
google-auth==1.11.3
grpcio==1.27.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kubernetes==11.0.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
python-dateutil==2.8.1
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.0
wrapt==1.11.2
